+++ packages/network_info_plus/tizen/src/network_info_plus_tizen_plugin.cc
@@ -19,7 +19,7 @@
 
 class NetworkInfoPlusTizenPlugin : public flutter::Plugin {
  public:
-  enum WifiInfoType { ESSID, BSSID };
+  enum WifiInfoType { ESSID, BSSID, SUBNET_MASK, GATEWAY_ADDR };
 
   static void RegisterWithRegistrar(flutter::PluginRegistrar *registrar) {
     auto channel =
@@ -63,8 +63,18 @@ class NetworkInfoPlusTizenPlugin : public flutter::Plugin {
     if (errorcode == WIFI_MANAGER_ERROR_NONE && current_ap != nullptr) {
       if (type == WifiInfoType::ESSID) {
         errorcode = wifi_manager_ap_get_essid(current_ap, &name);
-      } else {
+      } else if (type == WifiInfoType::BSSID) {
         errorcode = wifi_manager_ap_get_bssid(current_ap, &name);
+      } else if (type == WifiInfoType::SUBNET_MASK) {
+        // The requested subnet mask is implicitly ipv4.
+        // https://github.com/fluttercommunity/plus_plugins/blob/bd0262e5f4627358bfb42481a84122f60921d98b/packages/network_info_plus/network_info_plus/android/src/main/java/dev/fluttercommunity/plus/network_info/NetworkInfo.java#L63
+        errorcode = wifi_manager_ap_get_subnet_mask(
+            current_ap, WIFI_MANAGER_ADDRESS_FAMILY_IPV4, &name);
+      } else {
+        // The requested gateway address is implicitly ipv4.
+        // https://github.com/fluttercommunity/plus_plugins/blob/bd0262e5f4627358bfb42481a84122f60921d98b/packages/network_info_plus/network_info_plus/android/src/main/java/dev/fluttercommunity/plus/network_info/NetworkInfo.java#L108
+        errorcode = wifi_manager_ap_get_gateway_address(
+            current_ap, WIFI_MANAGER_ADDRESS_FAMILY_IPV4, &name);
       }
       if (errorcode == WIFI_MANAGER_ERROR_NONE) {
         result = name;
@@ -91,11 +101,24 @@ class NetworkInfoPlusTizenPlugin : public flutter::Plugin {
       char *ip_addr = nullptr;
       if (connection_get_ip_address(connection_, CONNECTION_ADDRESS_FAMILY_IPV4,
                                     &ip_addr) != CONNECTION_ERROR_NONE) {
-        result->Error("-1", "Couldn't obtain current ip address");
+        result->Error("-1", "Couldn't obtain current ipv4 address");
         return;
       }
       replay = ip_addr;
       free(ip_addr);
+    } else if (method_call.method_name().compare("wifiIPv6Address") == 0) {
+      char *ip_addr = nullptr;
+      if (connection_get_ip_address(connection_, CONNECTION_ADDRESS_FAMILY_IPV6,
+                                    &ip_addr) != CONNECTION_ERROR_NONE) {
+        result->Error("-1", "Couldn't obtain current ipv6 address");
+        return;
+      }
+      replay = ip_addr;
+      free(ip_addr);
+    } else if (method_call.method_name().compare("wifiSubmask") == 0) {
+      replay = GetWifiInfo(WifiInfoType::SUBNET_MASK);
+    } else if (method_call.method_name().compare("wifiGatewayAddress") == 0) {
+      replay = GetWifiInfo(WifiInfoType::GATEWAY_ADDR);
     } else {
       result->NotImplemented();
       return;