+++ packages/network_info_plus/CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
-* Initial release
+* Initial release.
+
+## 1.1.0
+
+* Add IPv6 information.
+* Add gateway ip address information.
+* Add subnet mask information.